requirements.txt
requests
beautifulsoup4
tenacity
python-dotenv
pytest
pytest-cov
pytest-mock
